arol/requirements.txt
asgiref==3.5.1
astroid==2.11.5
backports.zoneinfo==0.2.1
black==22.3.0
click==8.1.3
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.5.1
Django==4.0.4
django-environ==0.8.1
django-filter==21.1
django-import-export==2.8.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
importlib-metadata==4.11.4
isort==5.10.1
lazy-object-proxy==1.7.1
Markdown==3.3.7
MarkupPy==1.14
mccabe==0.7.0
mypy-extensions==0.4.3
mysqlclient==2.1.0
odfpy==1.4.1
openpyxl==3.0.10
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
PyJWT==2.4.0
pylint==2.13.9
pytz==2022.1
PyYAML==6.0
reportlab==3.6.9
sqlparse==0.4.2
tablib==3.2.1
tomli==2.0.1
typing_extensions==4.2.0
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.0
